updates listing table to store lat and lng as floats

+++ db/migrate/20210413110618_add_latitude_and_longitude_to_listings.rb
@@ -1,6 +1,6 @@
 class AddLatitudeAndLongitudeToListings < ActiveRecord::Migration[6.1]
   def change
-    add_column :listings, :latitude, :decimal
-    add_column :listings, :longitude, :decimal
+    add_column :listings, :latitude, :float
+    add_column :listings, :longitude, :float
   end
 end
